issue #141: moved SmartHomeObjectMapper to :app and removed Jackson from :core main sources
